Split up the EventAlerts file
